Fixing an issue with environments
When calling the launchdarkly api endpoint to fetch all of the feature
flags for a project, the response does not include environment data.

Calling the api endpoint for each flag does include environment data.

This did make me realise that each project can have diffferent
environments, so I moved the environment data to the transform step
of the etl process.

This feels more appropriate anyway as the transform is not not dependent
on the information coming into it, and operates only on transforming
the feature flags.